Add files via upload
New static module. Intended primarily for the sidebar, as links. The name and Twig template can be edited. Similarly, a Blueprint needs to be created. Reason: I need to add additional text information to the sidebar for which links are not suitable. On the page https://www.nocliteratury.tu.cz/historie/2024/mista, a static module for ROZCESTNÍK is used, into which I insert content via the Page Inject Plugin.